# docs/swagger.yaml
basePath: /
definitions:
  models.Song:
    properties:
      artist:
        type: string
      id:
        type: integer
      title:
        type: string
    type: object
  models.SongRequest:
    properties:
      artist:
        type: string
      title:
        type: string
    type: object
  utils.ErrorResponse:
    properties:
      message:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@songlibrary.com
    name: API Support
    url: http://www.songlibrary.com/support
  description: API для управления музыкальной библиотекой
  title: Music Library API
  version: "1.0"
paths:
  /songs:
    get:
      consumes:
      - application/json
      description: Fetch all the songs from the music library
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Limit of songs per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Song'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      summary: Get all songs
      tags:
      - songs
    post:
      consumes:
      - application/json
      description: Add a new song to the music library
      parameters:
      - description: Song details
        in: body
        name: song
        required: true
        schema:
          $ref: '#/definitions/models.SongRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Song'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      summary: Add a new song
      tags:
      - songs
swagger: "2.0"